Put website code into its own folder
