--- src/ts/CodeBlock.tsx
@@ -1,59 +1,0 @@
-import * as React from "react";
-import styled from "styled-components";
-import Highlight, { defaultProps, PrismTheme } from "prism-react-renderer";
-
-const Pre = styled.pre`
-  text-align: left;
-  margin: 1em 0;
-  padding: 0.5em;
-  overflow: scroll;
-
-  & .token-line {
-    line-height: 1.3em;
-    height: 1.3em;
-  }
-`;
-
-const Line = styled.div`
-  display: table-row;
-`;
-
-const LineNo = styled.span`
-  display: table-cell;
-  text-align: right;
-  padding-right: 1em;
-  user-select: none;
-  opacity: 0.5;
-`;
-
-const LineContent = styled.span`
-  display: table-cell;
-`;
-
-interface Props {
-  theme: PrismTheme;
-  code: string;
-}
-
-const CodeBlock: React.FunctionComponent<Props> = ({ theme, code }: Props) => {
-  return (
-    <Highlight {...defaultProps} theme={theme} code={code} language="cpp">
-      {({ className, style, tokens, getLineProps, getTokenProps }) => (
-        <Pre className={className} style={style}>
-          {tokens.map((line, i) => (
-            <Line key={i} {...getLineProps({ line, key: i })}>
-              <LineNo>{i + 1}</LineNo>
-              <LineContent>
-                {line.map((token, key) => (
-                  <span key={key} {...getTokenProps({ token, key })} />
-                ))}
-              </LineContent>
-            </Line>
-          ))}
-        </Pre>
-      )}
-    </Highlight>
-  );
-};
-
-export default CodeBlock;
